% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{checkCommonVariables}
\alias{checkCommonVariables}
\title{A function that takes a list of data frames and a list of common variables and checks that the common variables in each each dataset are of the same type.}
\usage{
checkCommonVariables(datasets, commonVariable)
}
\arguments{
\item{...}{List of data frames and common variables.}
}
\description{
This function takes a list of N data frames and common variables.
}
\keyword{data,variable,common}
\keyword{getDatasetList}

