% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{getCommonDataVariables}
\alias{getCommonDataVariables}
\title{A function for getting common data variables of N lists}
\usage{
getCommonDataVariables(...)
}
\arguments{
\item{...}{Lists passed as parameters}
}
\description{
This function takes N lists and returns their data variable(s)
}
\keyword{common,intersect,data,variable}

