% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{getCSVData}
\alias{getCSVData}
\title{A CSV Data Retrieval Function}
\usage{
getCSVData(dataURL)
}
\arguments{
\item{dataURL}{URL of CSV data}
}
\description{
This function allows you to retrieve CSV data at a specified URL in data frame format.
}
\examples{
getCSVData()
}
\keyword{CSV,}
\keyword{csv}
