% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{getDataset}
\alias{getDataset}
\title{A JSON & CSV Data Retrieval Function}
\usage{
getDataset(dataURL)
}
\arguments{
\item{dataURL}{of JSON or CSV data}
}
\description{
This function allows you to retrieve JSON or CSV data from a specified URL in data frame format.
}
\keyword{CSV,}
\keyword{csv,}
\keyword{json,}
\keyword{jsonstat}

