% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{getDatasetList}
\alias{getDatasetList}
\title{A function that takes a list of data frames and urls and replaces all urls with the corresponding data frame.}
\usage{
getDatasetList(datasets)
}
\arguments{
\item{...}{List of urls and data frames}
}
\description{
This function takes a list of N data frames and urls and returns a list of data frames.
}
\keyword{data,variable}
\keyword{getDatasetList}

