% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{getIntersectOfMany}
\alias{getIntersectOfMany}
\title{A function for getting common intersection of N lists}
\usage{
getIntersectOfMany(...)
}
\arguments{
\item{...}{Lists passed as parameters}
}
\description{
This function takes N lists and returns their common element(s)
}
\examples{
getIntersectOfMany()
}
\keyword{common,intersect}

