% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieval.R
\name{getJSONStatData}
\alias{getJSONStatData}
\title{A JSON Stat Data Retrieval Function}
\usage{
getJSONStatData(dataURL)
}
\arguments{
\item{dataURL}{URL of JSON-stat data}
}
\description{
This function allows you to retrieve JSON-stat data at a specified URL in data frame format.
}
\examples{
getJSONStatData()
}
\keyword{json,jsonstat}

