% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{getVariables}
\alias{getVariables}
\title{A function for presenting the data variables of N data sets}
\usage{
getVariables(...)
}
\arguments{
\item{...}{Lists passed as parameters}
}
\description{
This function takes N data sets and returns the different variables present in the data sets.
}
\keyword{getVariables}
\keyword{union,data,variable}

