% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R, R/data_retrieval.R,
%   R/user_functions.R
\docType{package}
\name{opendatr}
\alias{opendatr}
\alias{opendatr-package}
\alias{opendatr}
\alias{opendatr-package}
\alias{opendatr}
\alias{opendatr-package}
\title{Functions for manipulating datasets}
\description{
Functions for manipulating datasets

Functions for retrieving datasets

User functions for working with datasets
}

