% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plotting.R
\name{plotDataset}
\alias{plotDataset}
\title{A function that plots data based on the x and y values passed in}
\usage{
plotDataset(plotInfo)
}
\arguments{
\item{datasets}{datasets = [dataset0 = [xVals, yVals, xLabel, yLabel], dataset1 = [xVals, yVals, xLabel, yLabel], ..... ]}

\item{plots}{list of plots wanted (l = line, h = bar etc)}
}
\description{
A function that plots data based on the x and y values passed in
}
\keyword{data,plot}
\keyword{plotDataset}

