% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{plotDatasets}
\alias{plotDatasets}
\title{A function that plots N datasets on graph(s) based on the information entered by the user
parameters:}
\usage{
plotDatasets(datasets, graphTypes, commonVariable, yAxis, factors, xLabels,
  yLabels, titles)
}
\arguments{
\item{...}{Parameters in above format specifying the plot the user would like}
}
\description{
datasets = list of N data frames and urls
}
\details{
graphTypes = list of N character strings specfiying what graph to plot for each dataset, respectively.
"pie" for pie charts and "scatter" for scatter plots.

commonVariable = a character string common to all datasets or a list of N character strings, each
correspoinding to a dataset, respectively, specifying that they are to be considered common variables.

yAxis = list of N character strings specifying the values for the yAxis for each dataset, respectively.

factors = list of N lists. Each list corresponds to a dataset from the list of datasets, respectively.
In each list, an element name and value correspond to a factor column name and corresponding value
for that factor column, respectively.

xLabels = list of N character strings, specifying the label for the x axis for the corresponding graph

yLabels = list of N character strings, specifying the label for the y axis for the corresponding graph

titles = list of N character strings, specifying the label for the title for the corresponding graph
}
\keyword{common,intersect,data,variable}

