% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{getCommonDataVariables}
\alias{getCommonDataVariables}
\title{A function for getting common data variables of N data sets}
\usage{
getCommonDataVariables(datasets)
}
\arguments{
\item{...}{Lists passed as parameters}
}
\description{
This function takes N data sets and returns their common data variable(s)
}
\keyword{common,intersect,data,variable}

