% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{getCommonDataVariables}
\alias{getCommonDataVariables}
\title{A function for getting common data variables from N data sets}
\usage{
getCommonDataVariables(...)
}
\arguments{
\item{dataURL}{URL of JSON-stat data}
}
\description{
This function takes N data sets and returns their common data variables
}
\examples{
getCommonDataVariables()
}
\keyword{json,jsonstat}

