% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{getDatasetInfo}
\alias{getDatasetInfo}
\title{A function that returns plotting information about datasets}
\usage{
getDatasetInfo(...)
}
\arguments{
\item{...}{Lists passed as parameters}
}
\description{
This function takes N data sets as urls and returns a list of information about them
}
\keyword{common,intersect,data,variable}
