% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plotting.R
\name{plotDataSet}
\alias{plotDataSet}
\title{A function that plots data based on the x and y values passed in}
\usage{
plotDataSet(...)
}
\arguments{
\item{x}{x value to be plotted}

\item{y}{y value to be plotted}
}
\description{
A function that plots data based on the x and y values passed in
}
\keyword{data,plot}
