% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plotting.R
\name{plotDataSet}
\alias{plotDataSet}
\title{A function that plots data based on the x and y values passed in}
\usage{
plotDataSet(x, y, xlabel, ylabel, plotType)
}
\arguments{
\item{x}{x value to be plotted}

\item{y}{y value to be plotted}

\item{xlabel}{label for x axis}

\item{ylabel}{label for y axis}

\item{plotType}{Type of plot wanted (l = line, h = bar etc)}
}
\description{
A function that plots data based on the x and y values passed in
}
\keyword{data,plot}
