% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{plotDatasets}
\alias{plotDatasets}
\title{A function that plots N datasets on graph(s) based on the information entered by the user}
\usage{
plotDatasets(urls, graphType, commonVariable, yAxis, factors, xLabels, yLabels,
  titles)
}
\arguments{
\item{...}{List in above format specifying the plot the user would like}
}
\description{
This function takes a list of information in the specified format
list parameter =
[ urlList = [dataset0 = url0, dataset1 = url1, ...., datasetn = urln],
 graphType = [graph1, graph2, ..... graphn],
 commonVariable = var || [commonVar0, commonVar1, .... , commonVarn],
 yAxis = [dataset0 = colName, dataset1 = colName, ...., datasetN = colName],
 factors = [
   dataset0factors = [factor0 = val0, factor = val1, factor2 = val2, …., factorn = valn],
   dataset1factors = [factor0 = val0, factor1 = val1, factor2 = val2, …., factorn = valn],
   ....
   datasetNfactors = [factor0 = val0, factor1 = val1, factor2 = val2, …., factorn = valn]
   ]]
}
\keyword{common,intersect,data,variable}

